Fetch login person moved to auth provider
