+++ package.json
@@ -34,7 +34,7 @@
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@vitejs/plugin-react": "^4.0.3",
-    "rome": "12.1.3",
+    "rome": "^12.1.3",
     "typescript": "^5.2.2",
     "vite": "^4.4.5"
   }
